fix confirm order on exchange

+++ yii2/common/middleware/DataMiddleware.php
@@ -63,7 +63,7 @@ class DataMiddleware extends BaseObject
         return 0;
     }
 
-    public function getNeedBalance(): int
+    public function getNeedBalance(): float
     {
         return 0;
     }
